requirements.txt
ffmpeg-python==0.2.0
numpy==2.0.2
pillow==11.3.0